requirements.txt
google-genai
groq
SpeechRecognition
pyaudio
python-dotenv
edge-tts
pygame
keyboard
rich 